+++ inst/livy/spark-1.6.0/workerapply.scala
@@ -16,8 +16,9 @@ class WorkerApply(
   context: Array[Byte],
   options: Map[String, String],
   timeZoneId: String,
-  schema: org.apache.spark.sql.types.StructType
-  ) {
+  schema: org.apache.spark.sql.types.StructType,
+  genBarrierMap: () => Map[String, Any]
+  ) extends java.io.Serializable {
 
   import java.io.{File, FileWriter}
   import org.apache.spark._;
@@ -57,7 +58,8 @@ class WorkerApply(
       context,
       timeZoneId,
       schema,
-      options
+      options,
+      genBarrierMap()
     )
 
     val tracker = new JVMObjectTracker()