+++ inst/livy/spark-2.0.0/workerhelper.scala
@@ -108,7 +108,8 @@ object WorkerHelper {
         context,
         optionsImmMap,
         timeZoneId,
-        sourceSchema
+        sourceSchema,
+        () => Map()
       )
 
       workerApply.apply(rows)