+++ inst/livy/spark-2.4.0/workerapply.scala
@@ -1,6 +1,9 @@
-package sparklyr
+//
+// This file was automatically generated using livy_sources_refresh()
+// Changes to this file will be reverted.
+//
 
-class WorkerApply(
+class WorkerApply (
   closure: Array[Byte],
   columns: Array[String],
   config: String,
@@ -14,9 +17,10 @@ class WorkerApply(
   options: Map[String, String],
   timeZoneId: String,
   schema: org.apache.spark.sql.types.StructType,
-  genBarrierMap: () => Map[String, Any]
+  barrier: Boolean = false
   ) extends java.io.Serializable {
 
+  import java.io.Serializable;
   import java.io.{File, FileWriter}
   import org.apache.spark._;
 
@@ -44,6 +48,13 @@ class WorkerApply(
     // No point in starting up R process to not process anything
     if (!iterator.hasNext) return Array[org.apache.spark.sql.Row]().iterator
 
+    var barrierMap = Map[String, Any]()
+
+    if (barrier) {
+      val address = org.apache.spark.BarrierTaskContext.get().getTaskInfos().map(e => e.address)
+      barrierMap = Map("address" -> address)
+    }
+
     val workerContext = new WorkerContext(
       iterator,
       lock,
@@ -56,7 +67,7 @@ class WorkerApply(
       timeZoneId,
       schema,
       options,
-      genBarrierMap()
+      barrierMap
     )
 
     val tracker = new JVMObjectTracker()