+++ java/spark-2.0.0/workerhelper.scala
@@ -105,7 +105,8 @@ object WorkerHelper {
         context,
         optionsImmMap,
         timeZoneId,
-        sourceSchema
+        sourceSchema,
+        () => Map()
       )
 
       workerApply.apply(rows)